backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Data Validation (>=2.10 for Python 3.14 pre-built wheels; older 2.x may need Rust to build)
pydantic>=2.10.0,<3
pydantic-settings>=2.1.0
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# File Handling
aiofiles==23.2.1
Pillow>=10.2.0

# Utilities & Auth
PyJWT>=2.8.0
bcrypt>=4.0.0
python-dateutil==2.8.2

# AWS S3 (optional)
boto3==1.34.34

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Vector DB Support
pgvector>=0.2.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
#paddleocr>=2.7.0
numpy>=1.24.0
openai>=1.0.0
jsonschema>=4.19.0
gunicorn>=21.2.0
